Dirty fix for not consistent NMS archives
In the latest game updates its possible that some blocks are corrupt and not decompressable (gibberish blocks). Tested cases show that this is a thing on usually the last blocks of a file. Despite the fact that reported file size in the PAK is incosistent with the size of the decompressed stream, the output file looks ok! It seems like this issue is happening due to patchwork of the PAKs from previous versions. NOTE: Still not sure if this should be considered a permanent fix or not.